requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.3
astroid==2.3.3
backports.entry-points-selectable==1.1.0
colorama==0.4.3
distlib==0.3.3
Django==3.0.7
django-admin-interface==0.18.5
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-crispy-forms==1.8.1
django-filter==2.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-login-required-middleware==0.4
django-widget-tweaks==1.4.5
djangorestframework==3.13.1
filelock==3.3.2
future==0.18.2
html5lib==1.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pbr==5.8.0
Pillow==9.0.0
platformdirs==2.4.0
pylint==2.4.4
PyPDF2==1.26.0
python-barcode==0.13.1
python-bidi==0.4.2
pytz==2019.3
reportlab==3.6.5
six==1.13.0
sqlparse==0.3.0
stevedore==3.5.0
tzdata==2021.5
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.5
